major refactoring
Using SPKAC method now. The browser genererates the Pub/Priv key pair.
Send the Public signed key to the server at which point the server
issues a certificate signed with the CA and the SPKI public key it got from
the browser. The Certificate install URL is one time use only. Once the Client
Certificate is installed in the browser the user needs to backup the Certificate as it
cannot be reissued. If the user looses the Certificate a new one will be made.
On the install page the user must enter his username/password combination so the django auth
system can log him in and return his certificate. If the user already installed the certificate
or if there is not defined certificate for this user, and error is raised.